fix build issue

+++ player/lib/audio/portaudio/pa_host.h
@@ -49,10 +49,10 @@ extern "C"
 #endif
 
 #ifndef int32
-    typedef long int32;
+    typedef int int32;
 #endif
 #ifndef uint32
-    typedef unsigned long uint32;
+    typedef unsigned int uint32;
 #endif
 #ifndef int16
     typedef short int16;
